backend/requirements.txt
google.generativeai
youtube_transcript_api
flask
flask-cors